+++ package-utils.el
@@ -4,7 +4,6 @@
 ;; URL: https://github.com/Silex/package-utils
 ;; Keywords: package, convenience
 ;; Version: 0.4.2
-;; Package-Requires: ((async "1.6"))
 
 ;; This file is NOT part of GNU Emacs.
 
@@ -132,23 +131,6 @@ With prefix argument NO-FETCH, do not call `package-refresh-contents'."
                                   nil
                                   'require-match))))
   (package-delete (cadr (assoc name package-alist))))
-
-;;;###autoload
-(defun package-utils-install-async (package)
-  "Install PACKAGE asynchronously.
-
-Contrary to `package-install', PACKAGE can only be a symbol."
-  (interactive (list (car (eval (cadr (interactive-form 'package-install))))))
-  (require 'async)
-  (async-start
-   `(lambda ()
-      ,(async-inject-variables "^package-archives$")
-      ;; Initialize the package system if necessary.
-      (package-initialize t)
-      (package-install ',package))
-   `(lambda (result)
-      (package-initialize nil)
-      (message "%s installed" ',package))))
 
 (provide 'package-utils)